Add Docker build to Travis CI process
Motivation
----------
Ensure that the Docker build is successful on pull requests, and also
publish the image to Docker Hub on tags.

Modifications
-------------
Add Docker build to the build script.

Add a Docker publish process on tags to the deploy steps.

Results
-------
A Docker image is always built, but it is also published when tagged
using the tagged version as well as the ":latest" tag.

The deploy process expects the DOCKER_USERNAME and DOCKER_PASSWORD
environment variables to be present in the Travis configuration.  See
https://docs.travis-ci.com/user/docker/#Pushing-a-Docker-Image-to-a-Registry.